+++ backend/account/views.py
@@ -1,7 +1,6 @@
 import random
 
 from django.core.cache import cache
-from django.core.mail import send_mail
 from django.db import transaction
 from rest_framework import status
 from rest_framework.permissions import AllowAny
@@ -25,21 +24,8 @@ from account.serializers import (
     UserRegisterEmailVerificationSerializer,
     UserRegistrationSerializer,
 )
-from config.exceptions import (
-    EmailSendingFailedException,
-    InvalidRefreshToken,
-    TokenIssuanceException,
-)
-
-
-def send_auth_code_to_email(type: str, receiver: str, code: int) -> None:
-    try:
-        email_subject = "[ROCCIA 901] 본인 확인 인증 번호 입니다."
-        email_body = f"{type}을 위해 전송된 이메일입니다.\n 본인확인을 위해 인증 번호 [{code}]를 입력해 주세요."
-
-        send_mail(subject=email_subject, message=email_body, from_email="ROCCIA 901", recipient_list=[receiver])
-    except Exception as e:
-        raise EmailSendingFailedException()
+from account.tasks import send_auth_code_to_email
+from config.exceptions import InvalidRefreshToken, TokenIssuanceException
 
 
 class UserRegistrationAPIView(APIView):
@@ -120,7 +106,7 @@ class UserRegisterRequestAuthCodeAPIView(APIView):
         cache.set(f"{receiver}:register:code", code, timeout=600)
         cache.set(f"{receiver}:register:status", "uncertified", timeout=600)
 
-        send_auth_code_to_email("회원가입", receiver, code)
+        send_auth_code_to_email.delay("회원가입", receiver, code)
 
         return Response(
             # fmt: off
@@ -166,7 +152,7 @@ class PasswordUpdateRequestAuthCodeAPIView(APIView):
         cache.set(f"{receiver}:password:code", code, timeout=600)
         cache.set(f"{receiver}:password:status", "uncertified", timeout=600)
 
-        send_auth_code_to_email("비밀번호 변경", receiver, code)
+        send_auth_code_to_email.delay("비밀번호 변경", receiver, code)
 
         return Response(
             # fmt: off
